% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_downstream_gates.R
\name{assign_downstream_gates}
\alias{assign_downstream_gates}
\title{Assign Downstream Gates to Events from Single Cell Experiment Object}
\usage{
assign_downstream_gates(sce_object, gatingset, use_gate)
}
\arguments{
\item{sce_object}{A Single Cell Experiment object.}

\item{gatingset}{A GatingSet with GatingHierarchies.}

\item{use_gate}{A string defining which gate to use.}
}
\value{
A data frame with event names and assigned gates.
}
\description{
Assign Downstream Gates to Events from Single Cell Experiment Object
}
