% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_sce.R
\name{cluster_sce}
\alias{cluster_sce}
\title{Cluster and Run UMAP on Single Cell Experiment}
\usage{
cluster_sce(sce_object, use_class, seed = 1000, n = 5000)
}
\arguments{
\item{sce_object}{A Single Cell Experiment object.}

\item{use_class}{A string defining which class of markers to use: \code{"type"/"state"}}

\item{seed}{A numeric specifying which seed to use.}

\item{n}{A numeric specifying how many cells to subsample; NULL for all cells.}
}
\value{
A Single Cell Experiment object with clustering and UMAP.
}
\description{
Cluster and Run UMAP on Single Cell Experiment
}
