% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_panel.R
\name{create_panel}
\alias{create_panel}
\title{Create Panel Meta Data}
\usage{
create_panel(flowset)
}
\arguments{
\item{flowset}{flowSet containing FCS files}
}
\value{
data frame containing channel, antigen, and class information
}
\description{
Create Panel Meta Data
}
