% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_from_flowset.R
\name{.gate_data}
\alias{.gate_data}
\title{Apply Gate on GatingHierarchy}
\usage{
.gate_data(gating_hierarchy, gate_name)
}
\arguments{
\item{gating_hierarchy}{A GatingHierarchy.}

\item{gate_name}{A string with the name of the gate to be used.}
}
\value{
A data frame with the gated data.
}
\description{
Apply Gate on GatingHierarchy
}
