% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_from_flowset.R
\name{.transform_xpr}
\alias{.transform_xpr}
\title{Asinh Transformation of Expression Data}
\usage{
.transform_xpr(xpr_data, channels, cofactor = 150)
}
\arguments{
\item{xpr_data}{Expression data.}

\item{channels}{Paramters to have the transformation applied to.}

\item{cofactor}{Cofactor to use for transformation.}
}
\value{
Transformed expression data.
}
\description{
Asinh Transformation of Expression Data
}
