% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates_to_list.R
\name{gates_to_list}
\alias{gates_to_list}
\title{Title}
\usage{
gates_to_list(data, use_gates)
}
\arguments{
\item{data}{A data frame with event names and assigned gates.}

\item{use_gates}{A vector with gates to be used to create the list.}
}
\value{
A list with gates as names and containing event names.
}
\description{
Title
}
