% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowjo_gating_set.R
\name{get_gating_set}
\alias{get_gating_set}
\title{Apply FlowJo Gating on FlowSet}
\usage{
get_gating_set(flowset, flowjo_workspace)
}
\arguments{
\item{flowset}{A flowSet containing flowFrames.}

\item{flowjo_workspace}{FlowJo workspace file.}
}
\value{
A GatingSet containing GatingHierarchies.
}
\description{
Uses a FlowJo workspace file to apply gating, transformations, etc. on a flowSet.
}
