% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{list_fcs}
\alias{list_fcs}
\title{List all .FCS files in a given path}
\usage{
list_fcs(fcs_path)
}
\arguments{
\item{fcs_path}{A string with the path to the .FCS files.}
}
\value{
A vector with the .FCS file names.
}
\description{
List all .FCS files in a given path
}
