% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_FCS_with_EvID.R
\name{load_FCS_with_EvID}
\alias{load_FCS_with_EvID}
\title{Load FCS files and add Event ID}
\usage{
load_FCS_with_EvID(files)
}
\arguments{
\item{files}{vector containing FCS file names}
}
\value{
A flowSet
}
\description{
Adds a new parameter to the FCS file called "EvID" which contains unique event IDs for each event.
}
