% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{relevel_factor}
\alias{relevel_factor}
\title{Convert to Factor and Order Levels According to Template}
\usage{
relevel_factor(vect, level_order)
}
\arguments{
\item{vect}{A vector or factor.}

\item{level_order}{A vector with given order of the factor levels.}
}
\value{
A factor with the specified order of levels.
}
\description{
Convert to Factor and Order Levels According to Template
}
