% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_from_flowset.R
\name{sce_from_flowset}
\alias{sce_from_flowset}
\title{Create Single Cell Experiment Object from FlowSet}
\usage{
sce_from_flowset(
  flowset,
  use_gate,
  gatingset,
  sample_vector,
  sample_levels,
  panel_df
)
}
\arguments{
\item{flowset}{A flowSet object containing flowFrames.}

\item{use_gate}{A vector containing names of gates to use.}

\item{gatingset}{A GatingSet object to be used to derive Events in defined gates.}

\item{sample_vector}{A vector with sample names.}

\item{sample_levels}{A vector defining the levels of the sample conditions.}

\item{panel_df}{A data frame with the panel information.}
}
\value{
A Single Cell Experiment object with the gated FCS files.
}
\description{
Create Single Cell Experiment Object from FlowSet
}
