% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_gate_designation.R
\name{unique_gate_designation}
\alias{unique_gate_designation}
\title{Designate Non-Overlapping Gates to Column Data of a Single Cell Experiment}
\usage{
unique_gate_designation(sce_obj, gate_list, gate_levels, col_name)
}
\arguments{
\item{sce_obj}{A Single Cell Experiment Object.}

\item{gate_list}{A list with gates as names and containing event names.}

\item{gate_levels}{A vector containing the order of the gates.}

\item{col_name}{A string defining the column name in the colData of the Single Cell Experiment object.}
}
\value{
A Single Cell Experiment object with the added column containing the designated gates.
}
\description{
Designate Non-Overlapping Gates to Column Data of a Single Cell Experiment
}
